// src/app/logic/models/habitacion.model.ts
export class HabitacionModel{
  id: number;
  codigo: string = "";
  agencia_id: string = "";
  tipo_habitacion_id: string = "";
  tipo_habitacion: string = "";
  estado_habitacion_id: string = "";
  nro_habitacion: string = "";
  piso: string = "";
  descripcion: string = "";
  precio: number;
  color: string = "";
  estado_habitacion: string = "";
  color_estado: string = "";
}